api/requirements.txt
django~=5.1
djangorestframework~=3.15
gunicorn~=23.0